system_server: adjust swiftshader execmem special handling
As of bug 404926144, we are now running SELinux CTS neverallow tests
on all build types, including userdebug builds. This is problematic
for cuttlefish builds, which are now failing CTS due to the
workarounds added as part of bug 65201432.

Annotate the workaround neverallow block with an empty
userdebug_or_eng statement. This has the effect of emitting a
special marker which CTS uses to not run this test on userdebug
or eng devices.

Bug: 404926144
Bug: 406890511
Flag: TEST_ONLY
Test: atest CtsSecurityHostTestCases:android.security.cts.SELinuxNeverallowRulesTest
Change-Id: Ic23c48f7ec2f97e0bfdacf961b9c822355584a37