Export file_contexts_test_data to CTS
Instead of comparing the AOSP file_contexts with the device /system
file_contexts, reuse the file_contexts_test_data to make sure that all
security contexts are resolved similarly to AOSP.

When used with -t, checkfc verify that:
  1. All entries resolve to their expected context;
  2. All rules in file_contexts are exercised at least one.
For CTS, only the first check is relevant. Update checkfc to return a
different error code for the second verification.

Bug: 347112668
Test: atest CtsSecurityHostTestCases:android.security.cts.SELinuxHostTest#testAospFileContexts
Flag: TEST_ONLY
Change-Id: Ib7691451d0becd6728f44bf8664d6f5fb1feaeb5