microdroid: allow init to restorecon apex-info-list.xml
To reduce the boottime, restorecon is moved from apexd to init.  In this
change, init is allowed to getattr for the apex-info-list file. Note
that, in microdroid, it's still apexd to restorecon the apex-info-list
file. Hence when init tries to do it again, it getattr's the file first
and skips the operation.

Checked that apexd-bootstrap duration is reduced:
* cuttlefish: ~25ms (~13%)
* p24: ~60ms (~8%)

Bug: 414242227
Test: MicrodroidTestApp
Test: device boots
Change-Id: Ifc9650e90e2a56a02fc2600571a7fa70acd697b1