neverallow /proc/PID/fd directory/file ops
The /proc/self/fd/ directory contains open file descriptors.
The contents of this directory are managed by the kernel, and
userspace can't add or remove files from this directory.

Unfortunately, attempts to modify this directory will still
generate SELinux denials, even though the operation is
subsequently disallowed by the kernel. The easiest way to
reproduce this is to run:

  adb root
  adb shell
  setenforce 0
  runcon u:r:toolbox:s0 /system/bin/sh
  echo FOO > /proc/self/fd/BAR

Resulting in the following error message:

  # echo FOO > /proc/self/fd/BAR
  /system/bin/sh: can't create /proc/self/fd/BAR: No such file or directory

and the following SELinux denials:

  avc:  denied  { dac_override } for  comm="sh" capability=1  scontext=u:r:toolbox:s0 tcontext=u:r:toolbox:s0 tclass=capability permissive=1
  avc:  denied  { write } for  comm="sh" name="fd" dev="proc" ino=163777 scontext=u:r:toolbox:s0 tcontext=u:r:toolbox:s0 tclass=dir permissive=1
  avc:  denied  { add_name } for  comm="sh" name="BAR" scontext=u:r:toolbox:s0 tcontext=u:r:toolbox:s0 tclass=dir permissive=1
  avc:  denied  { create } for  comm="sh" name="BAR" scontext=u:r:toolbox:s0 tcontext=u:object_r:toolbox:s0 tclass=file permissive=1
  avc:  denied  { associate } for  comm="sh" name="BAR" scontext=u:object_r:toolbox:s0 tcontext=u:object_r:proc:s0 tclass=filesystem permissive=1

None of these denials make sense. They represent an artifact
of the kernel performing an access control check for an operation
which will eventually fail for other reasons.

With the exception of dac_override, disallow these kinds of
allow rules, to prevent others from blindly adding things to
policy which don't make sense. Fixup a few areas where
"associate" was incorrectly added to certain domains.

Test: compiles and boots
Bug: 394513737
Flag: EXEMPT bugfix / neverallow assertion
Change-Id: I805c5bcdc2612d0bd9229d82004db3a34237b11c