Revert "Add rules to cover memfd's for testing."
The use of memfd descriptors in an fexec() context
has similarities to https://g-issues.chromium.org/issues/40059061
which has been reported through ChromeOS VRP. As this
functionality is not used, remove the SELinux policy
which supports it.

The existing policy is also problematic by virtue
of being in a userdebug_or_eng block. Such blocks have
non-obvious differences in behavior which could only
reveal itself when running on -user builds, and as such,
should generally be avoided.

This reverts commit cda514a99b952012842ba55550b8934c8d92422a.

Flag: EXEMPT bugfix
Bug: 425672336
Test: compiles and boots
Change-Id: I71e213e63cfc9a68fce7df82508802a8fd0a0139