Allow virtualizationservice to talk to virtmgr
virtmgr is a per-app process where IVirtualMachine binder object is
located. virtualizationservice is the global service where each virtmgr
registers IVirtualMachine(s) onto.

Some debugging commands like `vm list` talks to virtualizationservice to
get information about all VMs in the system. virtualization then gathers
information by querying IVirtualMachines each of which is hosted in
virtmgr processes.

To support this use case, virtualizationservice needs to talk to virtmgr
over binder.

For now, this is only for debugging, but in a follow-up change, the
channel will be used to control the VMs; for example shutting them down
when the Android is about to shutdown or reboot.

Bug: 418877672
Flag: EXEMPT refactoring
Test: `vm list`
Running VMs: [
    VirtualMachineDebugInfo {
        name: "VmRunApp",
        cid: 2048,
        temporaryDirectory: "/data/misc/virtualizationservice/2048",
        requesterUid: 2000,
        requesterPid: 1577,
        hostConsoleName: None,
    },
]

Change-Id: Iddc9656739254370b75c81cc5a1ceb0a7864935f