Revert^2 "Add rules to cover memfd's for testing."
This reverts commit f53fd1f6af6643d25459579e811b1d9924c9bcef.

Reason for revert: The original commit which added memfd open
capabilities to the shell domain was in 2021. Since that time,
other teams have implicitly started relying on this, and the revert
of the 2021 change has broken their tests. Restore the functionality
for now, to get the tests green again and to give more time for an
analysis of the problem.

Please see bug 426059319 for a list of tests broken.

Bug: 426059319
Bug: 425672336
Change-Id: I13fd06b56d914343244dd24e504d2aa38b636703