tracefs: allow "f2fs_set_page_dirty" events on "user" builds
Without this change, the tracepoint is only accessible on debuggable
builds. This tracepoint has been proven to be useful in droidfood
tracing (as a proxy for flash wearout), and there is interest in making
it available on release (in particular, beta) builds.

The tracepoint contents aren't more sensitive than the
already-allowlisted f2fs events. Example of the event contents, in
human-readable format:
   dev = (254,60), ino = 11845, DATA, FILE, index = 0, dirty = 1, uptodate = 1
   dev = (254,60), ino = 1, NODE, FILE, index = 11845, dirty = 0, uptodate = 1

Tracepoint definition in the kernel:
  https://github.com/torvalds/linux/commit/26c6b88

Bug: 416499742
Change-Id: Id4a4a6d9c7707b65ee70091e9700874f8aa057f2