apexd: add ro.apexd.changed_active_apexes
This is a new property that's private to apexd. The property is to carry
the information from apexd-bootstrap to apexd.

Reason for ro. is to surpass the length limitation. The value of the
property is a concatenated list of APEX names, which can be longer than
the limit.

Bug: 386290694
Flag: com.android.apex.flags.mount_before_data
Test: ApexTestCases
Change-Id: I673e4532667529a5cc411ab4674832cd92c925e4