[nfc] Allow wakelock use
We are seeing reports of the device going to sleep during an active NFC transaction. This is causing race conditions with the NFC chip. To fix this, NFC stack needs to hold a native wakelock while sending data to the chip. See attached bug for more context.

Bug: 421823278
Test: SELinuxUncheckedDenialBootTest
Flag: EXEMPT minor change
Change-Id: I7e9ff0f9467a6039fc1374c196e7ec8dce1f0b61