Configure serial_service and native_serial_service
native_serial_service is an optional system service in AOSP to support
wired serial ports. Using a native service allows us to use
libueventd-rs which helps us parse sysfs metadata.

It is beneficial to define labels, assign the correct label to the
executable and service in AOSP because it not only benefits OEMs that
need it, but also ensure we use unified labels for them which allows us
to introduce neverallows should we encounter any needs. In fact it
already comes with a few neverallows because it is defined as a
coredomain.

Also set some access controls for serial_service and
native_serial_service.

Bug: 394262006
Test: native_serial_service is registered at boot.
Flag: android.hardware.serial.flags.enable_wired_serial_api
Change-Id: I4a1b56ce9d8b37a48d4adfbb6c0f8bea993e76e5